STM32: Start the DMA TX transfer after enabling the SPI peripheral.
When using a GPDMA transfer trigger, the GPDMA will memorize any trigger
events that occur between the start of the DMA transfer and the start of
the SPI transmission. This leads to additional transmissions in the
beginning. Moving the start of the DMA transfer after enabling the SPI
transmission fixes this.